Added WebPack, put the examples in a separate folder
This is a big commit making a big deal for us. First, our engine is now built into a single JavaScript file using Webpack and can be easily used via script-tag, CommonJS, etc. Second, I reorganized the project structure, put the examples in a separate folder and rewrote the code to depend on the Webpack output rather than the engine internals.